--- docs/index.html
@@ -1,30 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>To Do List</title>
-    <script src="https://kit.fontawesome.com/a70cf7601d.js" crossorigin="anonymous"></script>
-<script defer src="bundle.js"></script></head>
-
-<body>
-    <section class="todo">
-        <div class="flex-container">
-            <div class="todo-head flex-item">
-                <p class="title">Today's to dos</p>
-                <i class="fas fa-sync-alt"></i>
-            </div>
-            <div class="add-task">
-                <input class="task-input flex-item" type="text" name="task" id="task-input"
-                    placeholder="Add to your list...">
-                <p class="carriageReturn">&crarr;</p>
-            </div>
-            <ul class="todo-list flex-container">
-            </ul>
-            <button type="button" class="clear-completed flex-item">Clear all completed</button>
-        </div>
-    </section>
-</body>
-
-</html>
