--- src/__mocks__/getLS.js
@@ -1,8 +1,0 @@
-const getTasksLocalStorage = () => {
-    if (localStorage.getItem('tasksContainer') != null) {
-      return JSON.parse(localStorage.getItem('tasksContainer'));
-    }
-    return [];
-  };
-  
-  export default getTasksLocalStorage;
