+++ src/__mocks__/getLocalStorage.js
@@ -1,0 +1,8 @@
+const getTasksLocalStorage = () => {
+  if (localStorage.getItem('tasksContainer') != null) {
+    return JSON.parse(localStorage.getItem('tasksContainer'));
+  }
+  return [];
+};
+
+export default getTasksLocalStorage;
