+++ src/__mocks__/setLocalStorage.js
@@ -1,0 +1,5 @@
+const setTasksLocalStorage = (tasksContainer) => {
+  localStorage.setItem('tasksContainer', JSON.stringify(tasksContainer));
+};
+
+export default setTasksLocalStorage;
