--- src/__mocks__/setLs.js
@@ -1,6 +1,0 @@
-const setTasksLocalStorage = (tasksContainer) => {
-    localStorage.setItem('tasksContainer', JSON.stringify(tasksContainer));
-  };
-  
-  export default setTasksLocalStorage;
-
